refactor(react-broadcast): inject-feature-toggles to be hook based
